v. 0.5.2 Refactored package structures, added StreamOps class and StreamStorageFactory class
