+++ src/main/java/com/nanosai/streamops/StreamOps.java
@@ -1,0 +1,18 @@
+package com.nanosai.streamops;
+
+import com.nanosai.streamops.iteration.RecordIterator;
+import com.nanosai.streamops.storage.StreamStorageFactory;
+
+public class StreamOps {
+
+
+    public static RecordIterator createRecordIterator(){
+        return new RecordIterator();
+    }
+
+    public static StreamStorageFactory createStreamStorageFactory() {
+        return new StreamStorageFactory();
+    }
+
+
+}
