+++ src/main/java/com/nanosai/streamops/storage/file/StreamStorageFS.java
@@ -1,4 +1,4 @@
-package com.nanosai.streamops.engine.storage.file;
+package com.nanosai.streamops.storage.file;
 
 import com.nanosai.streamops.rion.RionUtil;
