+++ src/test/java/com/nanosai/streamops/storage/file/StreamStorageFSTest.java
@@ -1,4 +1,4 @@
-package com.nanosai.streamops.engine.storage.file;
+package com.nanosai.streamops.storage.file;
 
 import org.junit.jupiter.api.Test;
