+++ src/test/java/com/nanosai/streamops/storage/file/index/StreamStorageIndexFSTest.java
@@ -1,4 +1,4 @@
-package com.nanosai.streamops.engine.storage.file.index;
+package com.nanosai.streamops.storage.file.index;
 
 import org.junit.jupiter.api.Test;
